배포 오류 푸시
