+++ src/app/search/fetchSearch.ts
@@ -1,0 +1,16 @@
+export async function fetchSearch(debounceValue: string, searchText: string) {
+  const API_SERVER = process.env.NEXT_PUBLIC_API_SERVER;
+  const CLIENT_ID = process.env.NEXT_PUBLIC_CLIENT_ID;
+  const url = `${API_SERVER}/products?keyword=${searchText}`;
+  const res = await fetch(url, {
+    headers: {
+      "Content-Type": "application/json",
+      "client-id": `${CLIENT_ID}`,
+    },
+  });
+  const resJson = await res.json();
+  if (!resJson.ok) {
+    throw new Error("error");
+  }
+  return resJson.item;
+}
